auto detect environment when installing jupyter (#6629)
* instead of asking the user to select an
installer, we now autodetect the
environment being used, and use that installer.

* added comments and a constant for a product name

* changed errorHandler to use the installer with most
priority, fixed channelManager so it returns the highest priority
installer, increased the priority of the conda installer to 1

* updated the condaInstaller unit tests and
changed channel manager to return all supported installers.

* Added InstrallerNames.ts with an enum and a Map to
help identify the installers used by the extension.
Also updated all installers to use it. And returned channel
Manager to work as it used to.

* removed installerNames.ts and added 'name' to the
IModuleInstaller, to be able to use it in code to identify
the different installers.