Factor out RawFileSystem. (#9499)
(for #8995)

This separates the low-level filesystem operations from the high-level "fs utils". The move to the VS Code FS API can then happen strictly inside RawFileSystem. Also, this is part of what we reverted in #8970.

Note that the bulk of the changes are in test code. Also, of the non-test code, the changes are either just moving things around (from FileSystem to RawFileSystem), or adding the new RawFileSystem class (and IRawFileSystem interface).