Reduce the number of build agents used by CI. (#10221)
* Drop extra jobs from the PR-validation pipeline.

This reduces the number of build agents we are using most frequently,
without sacrificing much coverage.  Note that this relies on 2 things:

* sufficient unit test coverage
* manual "full CI" runs when there may be OS-specific concerns

With this change:

* all tests get run on 3.x on linux (including smoke)
* for 2.7 only the unit and functional tests are run (and only on linux)
* on Windows and OSX, only functional and "single workspace" tests are run

The total number of test jobs (and hence agents) drops from 13 to 8.

* Drop extra jobs from the PR-merge pipeline.

This change relies on the nightly CI run covering the full matrix.

The total number of test jobs (and hence agents) drops from 39 to 24.

* Limit # of parallel jobs in each job matrix in the PR-merge and nightly pipelines.

This reduces the max number of agents used by the PR-merge pipeline to 8
(from 24).  For the nightly pipeline it goes down from 84 to 12.  Both
will take longer to complete, but that shouldn't be a major problem.

* Add a "manual build" CI pipeline for faster turnaround.

* Do not test internal tools on lower than Python 3.7.

* Drop the temporary workaround jobs.

* (again) Do not test internal tools on lower than Python 3.7.