Move remaining "virtual"-env-related code to the py-envs component tree. (#12516)
This change is part of the work to isolate a "component" for Python environments. The focus here is on pulling over the remaining significant sections of code belonging to the "discovery" portion of the py-envs component. There are a few pieces here and there but we'll get those pulled over in later phases.

For those most part this PR only moves code between files. In some cases this involves refactoring code so that part of it can be moved.