Use the component adapter in the extension. (#13869)
This allows us to start using the new discovery code in the extension. The key thing is to be careful not to regress in the available functionality. So for now we have disabled use of the component.  When we feel comfortable with it we can enable it by setting the default for `ComponentAdapter.enabled` to `true`.

This PR involves small fixes to a large number of files due to API changes. The adapter is actually used (injected) in the following files:

src/client/interpreter/interpreterService.ts
src/client/interpreter/helpers.ts
src/client/pythonEnvironments/discovery/locators/index.ts
src/client/pythonEnvironments/discovery/locators/services/condaService.ts
src/client/pythonEnvironments/discovery/locators/services/windowsStoreInterpreter.ts